Re-implement the EventHandlerRpcHandler
EventHandlerRpcHandler handles the ClientCallable
server method and is thus not purely for polymer.
The hanlder was returned and tests were parted
between polymer (EventHandler) and server (ClientCallable)

Fixed tests to be more robust and also to work on Windows.

Closes #9116
