+++ .gitignore
@@ -53,3 +53,4 @@ flow-tests/**/pnpm-lock.yaml
 yarn.lock
 
 flow-client/src/main/resources/META-INF/resources/frontend/FlowClient.js
+flow-tests/**/types.d.ts
