+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/TaskRunNpmInstallTest.java
@@ -186,7 +186,7 @@ public class TaskRunNpmInstallTest {
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
         Mockito.verify(logger).info(captor.capture(),
                 Mockito.matches(getToolName()),
-                Mockito.matches(nodeModules.getAbsolutePath()), Mockito.any(),
+                Mockito.matches(nodeModules.getAbsolutePath().replaceAll("\\\\", "\\\\\\\\")), Mockito.any(),
                 Mockito.matches(Constants.PACKAGE_JSON));
         Assert.assertEquals(
                 "Skipping `{} install` because the frontend packages are already installed in the folder '{}' and the hash in the file '{}' is the same as in '{}'",
