+++ flow-tests/pom.xml
@@ -42,11 +42,6 @@
         </dependency>
         <dependency>
             <groupId>com.vaadin</groupId>
-            <artifactId>flow-polymer-template</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.vaadin</groupId>
             <artifactId>flow-dnd</artifactId>
             <version>${project.version}</version>
         </dependency>
