--- flow-tests/test-ccdm/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "no-name",
-  "license": "UNLICENSED",
-  "vaadin": {
-    "dependencies": {
-      "@vaadin/router": "1.7.2",
-      "@polymer/polymer": "3.2.0"
-    },
-    "devDependencies": {
-      "compression-webpack-plugin": "3.1.0",
-      "raw-loader": "4.0.0",
-      "webpack-cli": "3.3.11",
-      "script-ext-html-webpack-plugin": "2.1.4",
-      "awesome-typescript-loader": "5.2.1",
-      "terser": "4.6.7",
-      "copy-webpack-plugin": "5.1.1",
-      "lit-element": "2.3.1",
-      "webpack": "4.42.0",
-      "progress-webpack-plugin": "0.0.24",
-      "html-webpack-plugin": "3.2.0",
-      "typescript": "3.8.3",
-      "webpack-merge": "4.2.2",
-      "webpack-dev-server": "3.10.3"
-    },
-    "hash": "ef51f922bf1bf051c5a797844cef73ec710990189f97e2dc15ee4c3123551de0"
-  },
-  "dependencies": {
-    "@vaadin/router": "1.7.2",
-    "@polymer/polymer": "3.2.0",
-    "@vaadin/flow-frontend": "./target/flow-frontend"
-  },
-  "devDependencies": {
-    "compression-webpack-plugin": "3.1.0",
-    "raw-loader": "4.0.0",
-    "webpack-cli": "3.3.11",
-    "script-ext-html-webpack-plugin": "2.1.4",
-    "awesome-typescript-loader": "5.2.1",
-    "terser": "4.6.7",
-    "copy-webpack-plugin": "5.1.1",
-    "lit-element": "2.3.1",
-    "webpack": "4.42.0",
-    "progress-webpack-plugin": "0.0.24",
-    "html-webpack-plugin": "3.2.0",
-    "typescript": "3.8.3",
-    "webpack-merge": "4.2.2",
-    "webpack-dev-server": "3.10.3"
-  }
-}
