+++ flow-tests/test-embedding/embedding-test-assets/src/main/java/com/vaadin/flow/webcomponent/ThemedComponent.java
@@ -15,13 +15,12 @@
  */
 package com.vaadin.flow.webcomponent;
 
+import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.dependency.JsModule;
-import com.vaadin.flow.component.polymertemplate.PolymerTemplate;
-import com.vaadin.flow.templatemodel.TemplateModel;
 
 @JsModule("./src/themed-component.js")
 @Tag("themed-component")
-public class ThemedComponent extends PolymerTemplate<TemplateModel> {
+public class ThemedComponent extends Component {
 
 }
