+++ flow-tests/test-mixed/pom-npm-production.xml
@@ -22,6 +22,11 @@
             <artifactId>flow-test-resources</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>com.vaadin</groupId>
+            <artifactId>flow-polymer-template</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
