+++ flow-tests/test-multi-war/deployment/pom.xml
@@ -33,11 +33,11 @@
                     </webApp>
                     <contextHandlers>
                         <contextHandler implementation="org.eclipse.jetty.maven.plugin.JettyWebAppContext">
-                            <war>${project.basedir}/../test-war1/target/flow-test-multi-war1.war</war>
+                            <war>${project.basedir}${file.separator}..${file.separator}test-war1${file.separator}target${file.separator}flow-test-multi-war1.war</war>
                             <contextPath>/test-war1</contextPath>
                         </contextHandler>
                         <contextHandler implementation="org.eclipse.jetty.maven.plugin.JettyWebAppContext">
-                            <war>${project.basedir}/../test-war2/target/flow-test-multi-war2.war</war>
+                            <war>${project.basedir}${file.separator}..${file.separator}test-war2${file.separator}target${file.separator}flow-test-multi-war2.war</war>
                             <contextPath>/test-war2</contextPath>
                         </contextHandler>
                     </contextHandlers>
