+++ flow-tests/test-npm-only-features/test-npm-bytecode-scanning/src/main/java/com/vaadin/flow/testnpmonlyfeatures/bytecodescanning/MyButton.java
@@ -15,16 +15,13 @@
  */
 package com.vaadin.flow.testnpmonlyfeatures.bytecodescanning;
 
+import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.dependency.JsModule;
-import com.vaadin.flow.component.polymertemplate.PolymerTemplate;
-import com.vaadin.flow.templatemodel.TemplateModel;
 
 @Tag("my-button")
 @JsModule("./my-button.js")
-public class MyButton extends PolymerTemplate<MyButton.MyButtonModel> {
+public class MyButton extends Component {
 
-    public interface MyButtonModel extends TemplateModel {
 
-    }
 }
